Add support for symbol packages
