+++ src/Microsoft.DotNet.Tools.Pack/project.json
@@ -24,8 +24,8 @@
   },
   "scripts": {
     "postcompile": [
-       "../../scripts/build/place-binary \"%compile:OutputDir%/%project:Name%.dll\"",
-       "../../scripts/build/place-binary \"%compile:OutputDir%/%project:Name%.pdb\""
+      "../../scripts/build/place-binary \"%compile:OutputDir%/%project:Name%.dll\"",
+      "../../scripts/build/place-binary \"%compile:OutputDir%/%project:Name%.pdb\""
     ]
   }
 }
